fix: Remove API breaking change and introduce ContainerRegistryNodeV2 mutations
